#645 part 11: Moved titles, redirects, oEmbed, and list table methods.
Cleaned up code and lost track of time.

+++ bootstrap/plugin-activation.php
@@ -41,7 +41,7 @@ turn_on_autoloading: if ( ! is_headless( 'settings' ) ) {
 	$temp_options = $options;
 	// Write a small difference, so the change will be forwarded to the database.
 	if ( \is_array( $temp_options ) )
-		$temp_options['update_buster'] = (int) time();
+		$temp_options['update_buster'] = time();
 
 	$_success = \update_option( $setting, $temp_options, 'yes' );
 	if ( $_success )
